externalize common code into new trait
